+++ tests/test_model_methods.py
@@ -20,9 +20,9 @@ from tortoise.exceptions import (
     DoesNotExist,
     IntegrityError,
     MultipleObjectsReturned,
+    ObjectDoesNotExistError,
     OperationalError,
     ParamsError,
-    ObjectDoesNotExistError,
     ValidationError,
 )
 from tortoise.expressions import F, Q
@@ -90,9 +90,9 @@ class TestModelCreate(test.TestCase):
 class TestModelMethods(test.TestCase):
     async def asyncSetUp(self):
         await super().asyncSetUp()
-        self.mdl = await Tournament.create(name="Test")
-        self.mdl2 = Tournament(name="Test")
         self.cls = Tournament
+        self.mdl = await self.cls.create(name="Test")
+        self.mdl2 = self.cls(name="Test")
 
     async def test_save(self):
         oldid = self.mdl.id
@@ -175,6 +175,36 @@ class TestModelMethods(test.TestCase):
         self.assertNotEqual(self.mdl, mdl)
         mdl2 = await self.cls.get(name="Test2")
         self.assertEqual(mdl, mdl2)
+
+    async def test_update_or_create_with_defaults(self):
+        mdl = await self.cls.get(name=self.mdl.name)
+        mdl_dict = dict(mdl)
+        oldid = mdl.id
+        mdl.id = 135
+        with self.assertRaisesRegex(ParamsError, "Conflict value with key='id':"):
+            # Missing query: check conflict with kwargs and defaults before create
+            await self.cls.update_or_create(id=mdl.id, defaults=mdl_dict)
+        desc = str(uuid4())
+        # If there is no conflict with defaults and kwargs, it will be success to update or create
+        defaults = dict(mdl_dict, desc=desc)
+        kwargs = {"id": defaults["id"], "name": defaults["name"]}
+        mdl, created = await self.cls.update_or_create(defaults, **kwargs)
+        self.assertFalse(created)
+        self.assertEqual(defaults["desc"], mdl.desc)
+        self.assertNotEqual(self.mdl.desc, mdl.desc)
+        # Hint query: use defauts to update without checking conflict
+        mdl2, created = await self.cls.update_or_create(
+            id=oldid, desc=desc, defaults=dict(mdl_dict, desc="new desc")
+        )
+        self.assertFalse(created)
+        self.assertNotEqual(dict(mdl), dict(mdl2))
+        # Missing query: success to create if no conflict
+        not_exist_name = str(uuid4())
+        no_conflict_defaults = {"name": not_exist_name, "desc": desc}
+        no_conflict_kwargs = {"name": not_exist_name}
+        mdl, created = await self.cls.update_or_create(no_conflict_defaults, **no_conflict_kwargs)
+        self.assertTrue(created)
+        self.assertEqual(not_exist_name, mdl.name)
 
     async def test_first(self):
         mdl = await self.cls.first()