+++ tortoise/models.py
@@ -30,9 +30,9 @@ from tortoise.exceptions import (
     DoesNotExist,
     IncompleteInstanceError,
     IntegrityError,
+    ObjectDoesNotExistError,
     OperationalError,
     ParamsError,
-    ObjectDoesNotExistError,
 )
 from tortoise.fields.base import Field
 from tortoise.fields.data import IntField
@@ -1026,6 +1026,8 @@ class Model(metaclass=ModelMeta):
         :param using_db: Specific DB connection to use instead of default bound
         :param kwargs: Query parameters.
         :raises IntegrityError: If create failed
+        :raises TransactionManagementError: If transaction error
+        :raises ParamsError: If defaults conflict with kwargs
         """
         if not defaults:
             defaults = {}
@@ -1033,15 +1035,33 @@ class Model(metaclass=ModelMeta):
         try:
             return await cls.filter(**kwargs).using_db(db).get(), False
         except DoesNotExist:
+            return await cls._create_or_get(db, defaults, **kwargs)
+
+    @classmethod
+    async def _create_or_get(
+        cls, db: BaseDBAsyncClient, defaults: dict, **kwargs
+    ) -> Tuple[Self, bool]:
+        """Try to create, if fails with IntegrityError then try to get"""
+        for key in defaults.keys() & kwargs.keys():
+            if (default_value := defaults[key]) != (query_value := kwargs[key]):
+                raise ParamsError(f"Conflict value with {key=}: {default_value=} vs {query_value=}")
+        merged_defaults = {**kwargs, **defaults}
+        try:
+            async with in_transaction(connection_name=db.connection_name) as connection:
+                return await cls.create(using_db=connection, **merged_defaults), True
+        except IntegrityError as exc:
             try:
-                async with in_transaction(connection_name=db.connection_name) as connection:
-                    return await cls.create(using_db=connection, **defaults, **kwargs), True
-            except IntegrityError as exc:
-                try:
-                    return await cls.filter(**kwargs).using_db(db).get(), False
-                except DoesNotExist:
-                    pass
-                raise exc
+                return await cls.filter(**kwargs).using_db(db).get(), False
+            except DoesNotExist:
+                pass
+            raise exc
+
+    @classmethod
+    def _db_queryset(
+        cls, using_db: Optional[BaseDBAsyncClient] = None, for_write: bool = False
+    ) -> QuerySet[Self]:
+        db = using_db or cls._choose_db(for_write)
+        return cls._meta.manager.get_queryset().using_db(db)
 
     @classmethod
     def select_for_update(
@@ -1057,10 +1077,7 @@ class Model(metaclass=ModelMeta):
         Returns a queryset that will lock rows until the end of the transaction,
         generating a SELECT ... FOR UPDATE SQL statement on supported databases.
         """
-        db = using_db or cls._choose_db(True)
-        return (
-            cls._meta.manager.get_queryset().using_db(db).select_for_update(nowait, skip_locked, of)
-        )
+        return cls._db_queryset(using_db, for_write=True).select_for_update(nowait, skip_locked, of)
 
     @classmethod
     async def update_or_create(
@@ -1084,7 +1101,7 @@ class Model(metaclass=ModelMeta):
             if instance:
                 await instance.update_from_dict(defaults).save(using_db=connection)
                 return instance, False
-        return await cls.get_or_create(defaults, db, **kwargs)
+        return await cls._create_or_get(db, defaults, **kwargs)
 
     @classmethod
     async def create(
@@ -1141,10 +1158,7 @@ class Model(metaclass=ModelMeta):
         :param batch_size: How many objects are created in a single query
         :param using_db: Specific DB connection to use instead of default bound
         """
-        db = using_db or cls._choose_db(True)
-        return (
-            cls._meta.manager.get_queryset().using_db(db).bulk_update(objects, fields, batch_size)
-        )
+        return cls._db_queryset(using_db, for_write=True).bulk_update(objects, fields, batch_size)
 
     @classmethod
     async def in_bulk(
@@ -1161,8 +1175,7 @@ class Model(metaclass=ModelMeta):
         :param field_name: Must be a unique field
         :param using_db: Specific DB connection to use instead of default bound
         """
-        db = using_db or cls._choose_db()
-        return await cls._meta.manager.get_queryset().using_db(db).in_bulk(id_list, field_name)
+        return await cls._db_queryset(using_db).in_bulk(id_list, field_name)
 
     @classmethod
     def bulk_create(
@@ -1201,11 +1214,8 @@ class Model(metaclass=ModelMeta):
         :param batch_size: How many objects are created in a single query
         :param using_db: Specific DB connection to use instead of default bound
         """
-        db = using_db or cls._choose_db(True)
-        return (
-            cls._meta.manager.get_queryset()
-            .using_db(db)
-            .bulk_create(objects, batch_size, ignore_conflicts, update_fields, on_conflict)
+        return cls._db_queryset(using_db, for_write=True).bulk_create(
+            objects, batch_size, ignore_conflicts, update_fields, on_conflict
         )
 
     @classmethod
@@ -1213,8 +1223,7 @@ class Model(metaclass=ModelMeta):
         """
         Generates a QuerySet that returns the first record.
         """
-        db = using_db or cls._choose_db()
-        return cls._meta.manager.get_queryset().using_db(db).first()
+        return cls._db_queryset(using_db).first()
 
     @classmethod
     def filter(cls, *args: Q, **kwargs: Any) -> QuerySet[Self]:
@@ -1250,8 +1259,7 @@ class Model(metaclass=ModelMeta):
         """
         Returns the complete QuerySet.
         """
-        db = using_db or cls._choose_db()
-        return cls._meta.manager.get_queryset().using_db(db)
+        return cls._db_queryset(using_db)
 
     @classmethod
     def get(
@@ -1271,8 +1279,7 @@ class Model(metaclass=ModelMeta):
         :raises MultipleObjectsReturned: If provided search returned more than one object.
         :raises DoesNotExist: If object can not be found.
         """
-        db = using_db or cls._choose_db()
-        return cls._meta.manager.get_queryset().using_db(db).get(*args, **kwargs)
+        return cls._db_queryset(using_db).get(*args, **kwargs)
 
     @classmethod
     def raw(cls, sql: str, using_db: Optional[BaseDBAsyncClient] = None) -> "RawSQLQuery":
@@ -1286,8 +1293,7 @@ class Model(metaclass=ModelMeta):
         :param using_db: The specific DB connection to use
         :param sql: The raw sql.
         """
-        db = using_db or cls._choose_db()
-        return cls._meta.manager.get_queryset().using_db(db).raw(sql)
+        return cls._db_queryset(using_db).raw(sql)
 
     @classmethod
     def exists(
@@ -1304,8 +1310,7 @@ class Model(metaclass=ModelMeta):
         :param args: Q functions containing constraints. Will be AND'ed.
         :param kwargs: Simple filter constraints.
         """
-        db = using_db or cls._choose_db()
-        return cls._meta.manager.get_queryset().using_db(db).filter(*args, **kwargs).exists()
+        return cls._db_queryset(using_db).filter(*args, **kwargs).exists()
 
     @classmethod
     def get_or_none(
@@ -1322,8 +1327,7 @@ class Model(metaclass=ModelMeta):
         :param args: Q functions containing constraints. Will be AND'ed.
         :param kwargs: Simple filter constraints.
         """
-        db = using_db or cls._choose_db()
-        return cls._meta.manager.get_queryset().using_db(db).get_or_none(*args, **kwargs)
+        return cls._db_queryset(using_db).get_or_none(*args, **kwargs)
 
     @classmethod
     async def fetch_for_list(